Noodle Shop Website
Made this site for a hacker competition called a CTF (Capture The Flag) for my university, GCU. The purpose of this site is to allow cyber security students to learn and challenge themselves in an environment that allows them to test their skills without fear of persecution. It is responsive and uses CSS media queries as well. Languages used: HTML, CSS, and JavaScript.